+++ SnowboardCourses/src/main/java/ru/tsedrik/dao/PersonDAO.java
@@ -5,7 +5,7 @@ import ru.tsedrik.model.Role;
 
 import java.util.Collection;
 
-public interface PersonDAO extends GenericDAO<Person, Integer> {
+public interface PersonDAO extends GenericDAO<Person, Long> {
 
     Collection<Person> getAllByRole(Role role);
 }
