+++ SnowboardCourses/src/main/java/ru/tsedrik/service/PersonService.java
@@ -5,6 +5,6 @@ import ru.tsedrik.model.Person;
 public interface PersonService {
     void addPerson(Person person);
     Person deletePerson(Person person);
-    Person deletePersonById(int id);
-    Person getPersonById(int id);
+    Person deletePersonById(Long id);
+    Person getPersonById(Long id);
 }
